+++ python_scripts/translations.py
@@ -3,6 +3,7 @@ import json
 from natsort import natsorted
 
 translations = {}
+stellarity_translations = {}
 
 # in translations/enlightened
 for file in os.listdir("../translations/enlightened"):
@@ -11,17 +12,23 @@ for file in os.listdir("../translations/enlightened"):
 		data = json.load(f)
 		translations[file] = data
 
+for file in os.listdir("../translations/stellarity"):
+	print(file)
+	with open(f"../translations/stellarity/{file}", "r", encoding="utf-8") as f:
+		data = json.load(f)
+		stellarity_translations[file] = data
+
 done = False
 while not done:
 	print("Choose action:")
 	print("1. Add new key")
 	print("2. Remove key")
 	print("3. Rename key")
-	print("4. Add new key with custom value")
-	print("5. Exit (or just syncing from weblate)")
+	print("4. Copy key from stellarity")
+	print("x. Exit (or just syncing from weblate)")
 	action = input("Input action (1/2/3/4/5): ")
 
-	if action == "1" or action == "4":
+	if action == "1":
 		key = input("Enter new key: ")
 		value = input("Enter value for the new key: ")
 
@@ -36,8 +43,13 @@ while not done:
 		new_key = input("Enter new key name: ")
 
 		translations["en_us.json"][new_key] = translations["en_us.json"].pop(old_key)
+	elif action == "4":
+		key = input("Enter key to copy: ")
+		for stellarity_translation in stellarity_translations:
+			if key in stellarity_translations[stellarity_translation] and stellarity_translation in translations:
+				translations[stellarity_translation][key] = stellarity_translations[stellarity_translation][key]
 
-	elif action == "5":
+	elif action == "x":
 		done = True
 
 for lang in translations: